Write privileges weren't being enforced for processor-based writes
The edit check was looking for a global variable that wasn't set in the
processors. If someone knew what they were doing, they could modify (but not
read) data in a library they didn't have write access to. (To do this to a
specific library, they likely had to be a member, since I don't think the
libraryID is exposed except via classic sync.) The write would have been
recorded under the username of the modifying user.

API-based writes and syncs small enough to be processed inline were being
checked properly.