Add /keys/<key>, deprecating /users/<userID>/keys/<key>
And add a 'userID' property to the response.

Non-OAuth-based clients and tools can now use just the API key to get a
list of all libraries the key has access to (in combination with
/users/<userID>/groups, which may show public groups not listed in the
key's permissions). OAuth clients already had access to the userID from
the OAuth response, so this doesn't change much for them.

/users/<userID>/keys/<key> is now deprecated.

(This is API v3 only.)