libraryReading=members access fix, and config/setup changes for tests
Public group libraries with libraryReading=members were still accessible
to non-members via the API, though the library wasn't available on
www.zotero.org.

This also changes the way configuration is done for tests, moving to an
.ini file with settings that can be overridden at runtime.

The public groups necessary for the tests also now created
automatically, with other public groups owned by the primary test user
deleted. Private groups still have to be created manually and specified
in the config, but those will be switched to automatic creation
eventually.