Improve PUT/PATCH support and update version checking for writes
In API v3, it's now possible to PUT/PATCH to a missing object with a
given key, as long as version 0 is passed in the
If-Unmodified-Since-Version header or the JSON version property.
Previously, objects with client-generated keys had to be uploaded via
POST.

If a version other than 0 is included for a write for a missing object,
or if a version isn't provided, that's a 404 error. For multi-object
writes, specifying a version for a key that doesn't exist is a 404 for
that object.

For single-object writes, the header now must match the JSON property if
provided. If it doesn't, that's a 400 error. Previously the versions
were evaluated independently.