Don't send sort=itemType and sort=addedBy queries to read replicas
Those two sorts require table writes on the shards. This also adds a way
to avoid that for sort=itemType using ORDER BY FIELD(), and a similar
approach could be adopted for addedBy, but that could get problematic
for libraries with lots of items or custom item types. For now, just
keep those requests' queries on the writers.