Try to use DB writers last for read requests
When using Amazon Aurora, any node can be the writer or a reader at any
time, so each node's direct endpoint needs to be added to
shardHostReplicas. Replicas are chosen in a random order for each
request, but we want to try to send reads only to the readers. To
accomplish this, we check the status of innodb_read_only (which
indicates whether it's a writer) on the randomly chosen node, send it to
the end of the line if it's the writer, and make a note in memcached so
that other requests can do the same without checking the variable on the
node.