Update group metadata version on membership change [schema change]
This should cause permission changes to sync down to 5.0.

This also changes the group's version column to a larger integer. Since
there's no ?since for group metadata and the version is just a shorter
ETag, it previously wrapped around at 255. The syncing documentation
does technically say to check if the version number is different, not
higher, but since I didn't even code it that way in the client, we'll
just go with a normal monotonically incrementing version number.

ALTER TABLE `groups` CHANGE `version` `version` MEDIUMINT UNSIGNED NOT
NULL DEFAULT '1';