Update APIv3 dateModified handling (Zotero client only for now)
Previously, when an item was updated via the API, its `dateModified` was
updated with the current server timestamp if one of the following was
true:

1) `dateModified` wasn't provided
2) `dateModified` was provided but matched the existing timestamp

and also 3) properties other than `collections` and `relations` were
changed.

This commit changes the behavior for APIv3 for "Zotero" and "Firefox"
user agents to update the timestamp with the server time only if
`dateModified` isn't provided and properties other than `collections`,
`deleted`, `relations`, and `tags` are changed. Otherwise, the provided
`dateModified` value is used, even if it hasn't changed.

This gives clients better control over retaining the existing Date
Modified value. If a client wants to rely on the server timestamp, it
should unset `dateModified` in the upload.

This change is currently limited to the Zotero client, but it will be
applied to all APIv3 clients in the near future, so clients that depend
on the existing behavior should update their code to remove
`dateModified` on upload.