Don't update library version if uploaded settings are unchanged
Since settings are written in a single transaction, the library version
update on a write can be rolled back if nothing changed. Previously the
reverted library version was being returned but the updated version was
still being saved.

For other object types, multi-writes happen in separate transactions, so
a no-op still updates the library version (though this could possibly be
fixed).