Return 412 for nonexistent object if version > 0 is provided
Via If-Unmodified-Since-Version header or version property

Previously it returned 404, but 412 is in line with the spec for
If-Match: * when no entity exists. (It's not entirely clear what an
If-Unmodified-Since is supposed to return for a missing object, so
If-Match is the next best analogue.) And the implication for the client
is the same as for other object-level 412 responses: the local database
somehow fell out of sync with the current state, and it probably needs
to do a full sync.

See also: https://github.com/zotero/zotero/commit/41eb49cf7f98dfc4